refactor(refine): collapse the single-product loader factory
src/refine/loaders/index.ts dispatched on a ModelId union of exactly
one ('rmbg-1.4') and src/refine/loaders/types.ts modelled a generic
ModelLoader contract that no second model ever filled. The factory
was indirection without leverage — a tax every caller paid for an
abstraction that never branched.

Inline the types into rmbg14.ts (renamed ModelLoader → Rmbg14Loader,
dropped the redundant id field). ar-editor-advanced now imports
createRmbg14Loader directly. mobile-sam.ts is unaffected — it never
went through the factory; the SAM worker is its own contract.

If a second model ever lands, growing a registry from a single
concrete type is cheaper than the previous fake-generic boilerplate.
